(22/10) Proteção de rotas no back-end e no front-end

+++ back-end/src/controllers/users.js
@@ -7,14 +7,18 @@ const controller = {}     // Objeto vazio
 controller.create = async function(req, res) {
   try {
 
-      // Verifica se existe o campo "password" em "req.body"
-        // Caso positivo, geramos o hash da senha antes de enviá-la
-      // ao BD
-        // (12 na chamada a bcrypt.hash() corresponde ao número de
-        // passos de criptografia utilizados no processo)
-        if(req.body.password) {
-          req.body.password = await bcrypt.hash.apply(req.body.password, 12)
-        }
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
+    // Verifica se existe o campo "password" em "req.body".
+    // Caso positivo, geramos o hash da senha antes de enviá-la
+    // ao BD
+    // (12 na chamada a bcrypt.hash() corresponde ao número de
+    // passos de criptografia utilizados no processo)
+    if(req.body.password) {
+      req.body.password = await bcrypt.hash(req.body.password, 12)
+    }
 
     await prisma.user.create({ data: req.body })
 
@@ -31,10 +35,15 @@ controller.create = async function(req, res) {
 
 controller.retrieveAll = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+      
     const result = await prisma.user.findMany(
       // Omite o campo "password" do resultado
       // por questão de segurança
-      { omit: {password: true}}
+      { omit: { password: true } } 
     )
 
     // HTTP 200: OK (implícito)
@@ -50,10 +59,18 @@ controller.retrieveAll = async function(req, res) {
 
 controller.retrieveOne = async function(req, res) {
   try {
+
+    // Somente usuários administradores ou o próprio usuário
+    // autenticado podem acessar este recurso
+    // HTTP 403: Forbidden
+    if(! (req?.authUser?.is_admin || 
+      Number(req?.authUser?.id) === Number(req.params.id))) 
+      return res.status(403).end()
+      
     const result = await prisma.user.findUnique({
       // Omite o campo "password" do resultado
       // por questão de segurança
-      omit: { password: true},
+      omit: { password: true },
       where: { id: Number(req.params.id) }
     })
 
@@ -73,14 +90,19 @@ controller.retrieveOne = async function(req, res) {
 controller.update = async function(req, res) {
   try {
 
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     // Verifica se existe o campo "password" em "req.body".
-    // Caso positivo, greamos o hash da senha antes de enviá-la
+    // Caso positivo, geramos o hash da senha antes de enviá-la
     // ao BD
-      // (12 na chamada a bcrypt.hash() corresponde ao número de
+    // (12 na chamada a bcrypt.hash() corresponde ao número de
     // passos de criptografia utilizados no processo)
-    if (req.body.password) {
+    if(req.body.password) {
       req.body.password = await bcrypt.hash(req.body.password, 12)
     }
+
     const result = await prisma.user.update({
       where: { id: Number(req.params.id) },
       data: req.body
@@ -101,6 +123,11 @@ controller.update = async function(req, res) {
 
 controller.delete = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     await prisma.user.delete({
       where: { id: Number(req.params.id) }
     })
